+++ src/main/java/com/bobray/beernextdoor/controller/BreweryController.java
@@ -1,7 +1,9 @@
 package com.bobray.beernextdoor.controller;
 
 import com.bobray.beernextdoor.entity.Brewery;
+import com.bobray.beernextdoor.entity.User;
 import com.bobray.beernextdoor.repository.BreweryRepository;
+import com.bobray.beernextdoor.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -14,6 +16,9 @@ public class BreweryController {
     @Autowired
     BreweryRepository breweryRepository;
 
+    @Autowired
+    UserRepository userRepository;
+
     @GetMapping("/breweries")
     public List<Brewery> getAllBreweries() {
         return breweryRepository.findAll();
@@ -24,34 +29,55 @@ public class BreweryController {
 
         Optional<Brewery> breweryOptional = breweryRepository.findById(idBrewery);
         if (breweryOptional.isPresent()) {
+
             return breweryOptional.get();
         }
         return null;
     }
 
-    @PostMapping("/breweries")
-    public Brewery postBrewery(@RequestBody Brewery brewery) {
-        return breweryRepository.save(brewery);
-    }
+    @PostMapping("/{apiKey}/breweries")
+    public Brewery postBrewery(@PathVariable String apiKey,
+                               @RequestBody Brewery brewery) {
 
-    @PutMapping("/breweries/{idBrewery}")
-    public Brewery putBrewery(@PathVariable Long idBrewery,
-                              @RequestBody Brewery brewery) {
+        Optional<User> userOptional = userRepository.findUserByApiKey(apiKey);
+        if (userOptional.isPresent()) {
 
-        Optional<Brewery> breweryOptional = breweryRepository.findById(idBrewery);
-        if (breweryOptional.isPresent()) {
-            brewery.setIdBrewery(idBrewery);
             return breweryRepository.save(brewery);
         }
         return null;
     }
 
-    @DeleteMapping("/breweries/{idBrewery}")
-    public boolean deleteBrewery(@PathVariable Long idBrewery) {
-        Optional<Brewery> breweryOptional = breweryRepository.findById(idBrewery);
-        if (breweryOptional.isPresent()) {
-            breweryRepository.deleteById(idBrewery);
-            return true;
+    @PutMapping("/{apiKey}/breweries/{idBrewery}")
+    public Brewery putBrewery(@PathVariable String apiKey,
+                              @PathVariable Long idBrewery,
+                              @RequestBody Brewery brewery) {
+
+
+        Optional<User> userOptional = userRepository.findUserByApiKey(apiKey);
+        if (userOptional.isPresent()) {
+
+            Optional<Brewery> breweryOptional = breweryRepository.findById(idBrewery);
+            if (breweryOptional.isPresent()) {
+
+                brewery.setIdBrewery(idBrewery);
+                return breweryRepository.save(brewery);
+            }
+        }
+        return null;
+    }
+
+    @DeleteMapping("/{apiKey}/breweries/{idBrewery}")
+    public boolean deleteBrewery(@PathVariable String apiKey,
+                                 @PathVariable Long idBrewery) {
+
+        Optional<User> userOptional = userRepository.findUserByApiKey(apiKey);
+        if (userOptional.isPresent()) {
+
+            Optional<Brewery> breweryOptional = breweryRepository.findById(idBrewery);
+            if (breweryOptional.isPresent()) {
+                breweryRepository.deleteById(idBrewery);
+                return true;
+            }
         }
         return false;
     }
