+++ src/main/java/com/bobray/beernextdoor/controller/TypeController.java
@@ -1,7 +1,9 @@
 package com.bobray.beernextdoor.controller;
 
 import com.bobray.beernextdoor.entity.Type;
+import com.bobray.beernextdoor.entity.User;
 import com.bobray.beernextdoor.repository.TypeRepository;
+import com.bobray.beernextdoor.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -13,6 +15,9 @@ public class TypeController {
 
     @Autowired
     TypeRepository typeRepository;
+
+    @Autowired
+    UserRepository userRepository;
 
     @GetMapping("/types")
     public List<Type> getAllTypes() {
@@ -29,29 +34,49 @@ public class TypeController {
         return null;
     }
 
-    @PostMapping("/types")
-    public Type postType(@RequestBody Type type) {
-        return typeRepository.save(type);
-    }
+    @PostMapping("/{apiKey}/types")
+    public Type postType(@PathVariable String apiKey,
+                         @RequestBody Type type) {
 
-    @PutMapping("types/{idType}")
-    public Type putType(@PathVariable Long idType, @RequestBody Type type) {
+        Optional<User> userOptional = userRepository.findUserByApiKey(apiKey);
+        if (userOptional.isPresent()) {
 
-        Optional<Type> typeOptional = typeRepository.findById(idType);
-        if (typeOptional.isPresent()) {
-            type.setIdType(idType);
             return typeRepository.save(type);
         }
         return null;
     }
 
-    @DeleteMapping("/types/{idType}")
-    public boolean deleteType(@PathVariable Long idType) {
+    @PutMapping("/{apiKey}types/{idType}")
+    public Type putType(@PathVariable String apiKey,
+                        @PathVariable Long idType,
+                        @RequestBody Type type) {
 
-        Optional<Type> typeOptional = typeRepository.findById(idType);
-        if (typeOptional.isPresent()) {
-            typeRepository.deleteById(idType);
-            return true;
+        Optional<User> userOptional = userRepository.findUserByApiKey(apiKey);
+        if (userOptional.isPresent()) {
+
+            Optional<Type> typeOptional = typeRepository.findById(idType);
+            if (typeOptional.isPresent()) {
+
+                type.setIdType(idType);
+                return typeRepository.save(type);
+            }
+        }
+        return null;
+    }
+
+    @DeleteMapping("/{apiKey}/types/{idType}")
+    public boolean deleteType(@PathVariable String apiKey,
+                              @PathVariable Long idType) {
+
+        Optional<User> userOptional = userRepository.findUserByApiKey(apiKey);
+        if (userOptional.isPresent()) {
+
+            Optional<Type> typeOptional = typeRepository.findById(idType);
+            if (typeOptional.isPresent()) {
+
+                typeRepository.deleteById(idType);
+                return true;
+            }
         }
         return false;
     }
